+++ ChangeLog.md
@@ -2,8 +2,8 @@
 
 ## 1.5 - 2017-12-07
 ### Added
-- moon_phase_asfloat
 - this file
 
 ### Changed
 - dawn_utc, sunrise_utc, sunset_utc and dusk_utc now only raise AstralError for a math domain exception all other exceptions are passed through.
+- moon_phase now takes another parameter if the type to return either int (the default) or float